Refactor: 닉네임 중복검사 테스트 UserAcceptanceTest로 옮김
