+++ BE/house/src/test/java/com/nextsquad/house/acceptance/login/LoginAcceptanceTest.java
@@ -34,8 +34,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
-import static org.hamcrest.Matchers.*;
-
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.*;
 import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.documentationConfiguration;
@@ -108,40 +106,6 @@ public class LoginAcceptanceTest {
             .then()
                 .statusCode(HttpStatus.OK.value())
                 .log();
-    }
-
-    @Test
-    @DisplayName("이미 DB에 있는 닉네임을 넣고 중복검사를 요청하면 true가 응답된다")
-    void duplicateTrueTest() {
-        RestAssured
-            .given(spec)
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .filter(RestAssuredRestDocumentation.document("get-user-info", preprocessRequest(prettyPrint()), preprocessResponse(prettyPrint())))
-                .param("nickname", "lee")
-                .header("access-token", token.getAccessToken().getTokenCode())
-            .when()
-                .get("/users/check-duplication")
-            .then()
-                .statusCode(HttpStatus.OK.value())
-                .body("isDuplicated", is(true));
-    }
-
-    @Test
-    @DisplayName("DB에 없는 닉네임을 넣고 중복검사를 요청하면 false가 응답된다")
-    void duplicateFalseTest() {
-        RestAssured
-            .given(spec)
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .filter(RestAssuredRestDocumentation.document("get-user-info", preprocessRequest(prettyPrint()), preprocessResponse(prettyPrint())))
-                .param("nickname", "testnickname")
-                .header("access-token", token.getAccessToken().getTokenCode())
-            .when()
-                .get("/users/check-duplication")
-            .then()
-                .statusCode(HttpStatus.OK.value())
-                .body("isDuplicated", is(false));
     }
 
     @Test
